fix ts && lint
