+++ packages/cspell/src/lint.ts
@@ -14,6 +14,7 @@ import { buildGlobMatcher, extractGlobsFromMatcher, extractPatterns, normalizeGl
 import { measurePromise } from './util/timer';
 import * as util from './util/util';
 import { loadReporters, mergeReporters } from './util/reporters';
+import { toError } from './util/errors';
 
 export interface FileResult {
     fileInfo: FileInfo;
@@ -332,19 +333,4 @@ function getLoggerFromReporter(reporter: CSpellReporter): Logger {
         warn,
         error,
     };
-}
-
-function toError(e: unknown): Error {
-    if (isError(e)) return e;
-    return {
-        name: 'error',
-        message: format(e),
-    };
-}
-
-function isError(e: unknown): e is Error {
-    if (!e || typeof e !== 'object') return false;
-    if (e instanceof Error) return true;
-    const ex = <Error>e;
-    return typeof ex.name === 'string' && typeof ex.message === 'string';
 }
